conv_fe/requirements.txt
numpy~=1.19.5
pandas~=1.2.4
python-dateutil~=2.8.0
lightgbm~=2.3.0
scikit-learn~=0.21.3